Moved record dict functions to utils.py
Now that the record classes handle the SSIM data in their init,
move the functios that convert an SSIM into a dict into the utils file.
They can be useful for debugging, or looking at SSIM data - however I
don't seem to need them now that the record classes are more fleshed out.